[main] upd: signup, search, profile, premiere, person, mainpage in proccess
