+++ hw/code/locators/pagePersonLocators.py
@@ -1,0 +1,8 @@
+class PersonPageParams:
+    URL_PAGE = f'/person/31/'
+
+    # check redirect
+    CLASS_NAME_FILM = "film__blackout"
+
+    # check film contains person
+    CLASS_NAME_ROLE = "about-film__role"
