+++ hw/code/locators/pagePremiereLocators.py
@@ -1,0 +1,10 @@
+class PremierePageParams:
+    URL_PAGE = f'/premieres/'
+
+    # check premiere page's content
+    CLASS_NAME_PREMIERES_TITLE = 'premiere-page__title'
+    CLASS_NAME_PREMIERES_DAY = 'premiere-day'
+
+    # check film poster redirect
+    CLASS_NAME_PREMIERES_FILM_POSTER = 'premiere-film__poster-container'
+    CLASS_NAME_FILM_PAGE_ABOUT = 'js-film-page__about'
