+++ hw/code/locators/pageProfileLocators.py
@@ -1,0 +1,11 @@
+class ProfilePageParams:
+    URL_PAGE_PROFILE = '/profile/'
+    CLASS_CHANGE_SVG = 'profile__change__svg'
+    CLASS_FORM_CHANGE = 'profile__change'
+    CLASS_FORM_DISPLAY_NONE = 'dysplay-none'
+    CLASS_FORM_DISPLAY_FLEX = 'dysplay-flex'
+    CLASS_NAME_SAVE_NEW_NAME = 'profile__input__button'
+    CLASS_NAME_INFO_VALUE = 'profile__info__value'
+    RATES = 1
+    COLLS = 2
+    CLASS_NAME_FIELD_NAME = "profile__username"
