+++ hw/code/locators/pageSearchLocators.py
@@ -1,0 +1,9 @@
+class SearchPageParams:
+    URL_PAGE_SEARCH = "/"
+    CLASS_NAME_FIRST_FOUNDED_FILM = 'premiere-film__blackout'
+    CLASS_NAME_FIRST_FOUNDED_TITLE = 'premiere-film__information-title'
+    CLASS_NAME_TITLE_ON_FILM_PAGE = 'about-film__title'
+    CLASS_NAME_INPUT_SEARCH = 'js-header__form__input'
+    CLASS_NAME_BUTTON_RETURN_MAIN = 'search-page__no-content-btn'
+    CLASS_NAME_TITLE_SEARCH = 'search-page__title'
+    CLASS_NAME_CATEGORY_TITLE = 'search-list__title'
