+++ hw/code/locators/pageSignupLocators.py
@@ -1,0 +1,13 @@
+class SignupPageParams:
+    URL_PAGE_LOGIN = 'login'
+    URL_PAGE_SIGNUP = '/signup/'
+
+    CLASS_BUTTON_OPEN_LOGIN_WINDOW = 'modal__login__switch__btn'
+    CLASS_NAME_MODAL_INPUT = 'js-modal__input'
+    NICKNAME_INPUT = 0
+    EMAIL_INPUT = 1
+    PASSWORD_INPUT = 2
+    REPEAT_PASSWORD_INPUT = 3
+    CLASS_NAME_BUTTON_REG = "modal__input__button-auth"
+    CLASS_NAME_ALREADY_REGISTER = "js-modal__input__error"
+    CLASS_WRONG_INPUT = 'modal__input_red_border'
