+++ hw/code/test_film_auth.py
@@ -1,7 +1,5 @@
-from selenium.webdriver.common.by import By
 import unittest
 
-from utils.base_page import BasePage
 from utils.helper_auth import needed_auth
 from pages.commonPage import CommonPage
 from pages.pageFilm import FilmPage
