+++ hw/code/test_film_unauth.py
@@ -1,11 +1,7 @@
-from selenium.webdriver.common.by import By
 import unittest
 
-from utils.base_page import BasePage
 from pages.commonPage import CommonPage
 from pages.pageFilm import FilmPage
-from locators.pageFilmLocators import SelectorsFilm
-
 
 
 class TestFilmPageUnauthorized(unittest.TestCase, FilmPage):
